Update progress formatter to match changed file location
