+++ lib/reek/report/formatter.rb
@@ -2,6 +2,7 @@
 require_relative 'code_climate/code_climate_formatter'
 require_relative 'formatter/heading_formatter'
 require_relative 'formatter/location_formatter'
+require_relative 'formatter/progress_formatter'
 require_relative 'formatter/simple_warning_formatter'
 require_relative 'formatter/wiki_link_warning_formatter'
